feat(codex): migrate chrome-devtools MCP to direct stdio
Point codex's config.toml chrome-devtools entry at the SAME shared stdio
spec the Claude inject config uses (command + args from the browser install
module), dropping the supergateway :8767 bridge url. codex was put on the
bridge to contain a Linux orphan leak; cdp-mcp 1.1.1+ stdin-EOF shutdown
(+ the already-silenced warn flood) makes a direct stdio child die with the
codex session, so the bridge is no longer needed for either client.

Split the over-limit config generator (patch.py, 248 lines) into a
domain folder config-generator/: generate_config.py (orchestration) and
toml_render.py (pure TOML serialization), referenced from nix by directory
so the sibling module resolves from the store. Drop its line-count baseline.
