+++ home/base/codex/config-generator/generate_config.py
@@ -1,7 +1,10 @@
+import json
 import os
 import pathlib
 import tomllib
 from typing import Any
+
+from toml_render import render_codex_config_toml
 
 codex_default_model = os.environ.get("CODEX_DEFAULT_MODEL", "gpt-5.4")
 codex_developer_instructions = os.environ.get(
@@ -12,9 +15,8 @@ codex_developer_instructions = os.environ.get(
         "fast (default), deep, web."
     ),
 )
-chrome_devtools_mcp_streamable_http_bridge_url = os.environ[
-    "CODEX_CHROME_DEVTOOLS_MCP_STREAMABLE_HTTP_BRIDGE_URL"
-]
+chrome_devtools_mcp_command = os.environ["CODEX_CHROME_DEVTOOLS_MCP_COMMAND"]
+chrome_devtools_mcp_args = json.loads(os.environ["CODEX_CHROME_DEVTOOLS_MCP_ARGS_JSON"])
 
 
 def build_trusted_project_entries() -> dict[str, dict[str, str]]:
@@ -52,7 +54,8 @@ def build_trusted_project_entries() -> dict[str, dict[str, str]]:
 def build_mcp_server_entries() -> dict[str, dict[str, Any]]:
     return {
         "chrome-devtools": {
-            "url": chrome_devtools_mcp_streamable_http_bridge_url,
+            "command": chrome_devtools_mcp_command,
+            "args": chrome_devtools_mcp_args,
         },
     }
 
@@ -147,102 +150,4 @@ data.setdefault("mcp_servers", {})["chrome-devtools"] = build_mcp_server_entries
     "chrome-devtools"
 ]
 
-
-def toml_quote(value: str) -> str:
-    return '"' + value.replace("\\", "\\\\").replace('"', '\\"') + '"'
-
-
-def toml_value(value: Any) -> str:
-    if value is None:
-        return "null"
-    if isinstance(value, bool):
-        return "true" if value else "false"
-    if isinstance(value, int):
-        return str(value)
-    if isinstance(value, float):
-        return repr(value)
-    if isinstance(value, str):
-        return toml_quote(value)
-    if isinstance(value, list):
-        return "[" + ", ".join(toml_value(item) for item in value) + "]"
-    if isinstance(value, dict):
-        items = ", ".join(
-            f"{key} = {toml_value(nested_value)}"
-            for key, nested_value in sorted(value.items())
-        )
-        return "{ " + items + " }"
-    raise TypeError(type(value))
-
-
-def emit_table_header(parts: list[str]) -> str:
-    out = []
-    for part in parts:
-        if any(character in part for character in "/\\:") or part.startswith(
-            ("~", ".", " ")
-        ):
-            out.append(toml_quote(part))
-        else:
-            out.append(part)
-    return "[" + ".".join(out) + "]"
-
-
-NESTED_TABLES = {
-    "projects",
-    "profiles",
-    "features",
-    "analytics",
-    "mcp_servers",
-    "sandbox_workspace_write",
-    "tools",
-}
-
-lines = []
-
-for k in sorted(data.keys()):
-    if k in NESTED_TABLES or isinstance(data[k], dict):
-        continue
-    lines.append(f"{k} = {toml_value(data[k])}")
-
-for table_name in ["analytics", "tools", "sandbox_workspace_write", "features"]:
-    tbl = data.get(table_name)
-    if not isinstance(tbl, dict) or not tbl:
-        continue
-    lines.append("")
-    lines.append(f"[{table_name}]")
-    for k, v in sorted(tbl.items()):
-        lines.append(f"{k} = {toml_value(v)}")
-
-profiles = data.get("profiles")
-if isinstance(profiles, dict) and profiles:
-    for prof in sorted(profiles.keys()):
-        val = profiles[prof]
-        if not isinstance(val, dict):
-            continue
-        lines.append("")
-        lines.append(f"[profiles.{prof}]")
-        for k, v in sorted(val.items()):
-            lines.append(f"{k} = {toml_value(v)}")
-
-projects = data.get("projects")
-if isinstance(projects, dict) and projects:
-    for path in sorted(projects.keys()):
-        val = projects[path]
-        if not isinstance(val, dict):
-            continue
-        lines.append("")
-        lines.append(emit_table_header(["projects", path]))
-        for k, v in sorted(val.items()):
-            lines.append(f"{k} = {toml_value(v)}")
-
-mcp_servers = data.get("mcp_servers")
-if isinstance(mcp_servers, dict) and mcp_servers:
-    for name in sorted(mcp_servers.keys()):
-        val = mcp_servers[name]
-        if not isinstance(val, dict):
-            continue
-        lines.append("")
-        lines.append(f"[mcp_servers.{name}]")
-        for k, v in sorted(val.items()):
-            lines.append(f"{k} = {toml_value(v)}")
-
-config_path.write_text("\n".join(lines).rstrip() + "\n", encoding="utf-8")
+config_path.write_text(render_codex_config_toml(data), encoding="utf-8")
